pyproject.toml
[build-system]
requires = ["flit_core>=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aiida-submission-controller"
version = "0.1.2"
authors = [
  {name = "Giovanni Pizzi"},
  {name = "Austin Zadoks"},
  {name = "Marnik Bercx"}
]
description = "Utilities for controlling submission flow in AiiDA"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AiiDA",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]
requires-python = ">=3.6"

dependencies = [
    "aiida-core~=2.5",
    "rich",
]

[project.urls]
Source = "https://github.com/aiidateam/aiida-submission-controller"

[project.optional-dependencies]
qe = [
    "aiida-quantumespresso"
]
dev = [
    "pre-commit~=2.17.0",
    "pylint-pydantic~=0.1.8"
]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
